requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in requirements/requirements.testing.in
#
anyio==4.6.2.post1
    # via
    #   httpx
    #   jupyter-server
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==2.4.1
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
babel==2.16.0
    # via jupyterlab-server
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.2.0
    # via nbconvert
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.0
    # via requests
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.9
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==2.1.0
    # via stack-data
fastjsonschema==2.21.0
    # via nbformat
fonttools==4.55.0
    # via matplotlib
fqdn==1.5.1
    # via jsonschema
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.0
    # via jupyterlab
idna==3.10
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
ipykernel==6.29.5
    # via
    #   jupyter
    #   jupyter-console
    #   jupyterlab
ipython==8.29.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==8.1.5
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.2
    # via ipython
jinja2==3.1.4
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
json5==0.10.0
    # via jupyterlab-server
jsonpointer==3.0.0
    # via jsonschema
jsonschema[format-nongpl]==4.23.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2024.10.1
    # via jsonschema
jupyter==1.1.1
    # via -r requirements/requirements.in
jupyter-client==8.6.3
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.10.0
    # via jupyter-server
jupyter-lsp==2.2.5
    # via jupyterlab
jupyter-server==2.14.2
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.2.6
    # via
    #   jupyter
    #   notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.27.3
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.13
    # via ipywidgets
kiwisolver==1.4.7
    # via matplotlib
markupsafe==3.0.2
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.9.2
    # via -r requirements/requirements.in
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mistune==3.0.2
    # via nbconvert
nbclient==0.10.0
    # via nbconvert
nbconvert==7.16.4
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
notebook==7.2.2
    # via jupyter
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
numpy==2.1.3
    # via
    #   -r requirements/requirements.in
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   pyexr
    #   scipy
opencv-python==4.10.0.84
    # via -r requirements/requirements.in
openexr==3.3.2
    # via pyexr
overrides==7.7.0
    # via jupyter-server
packaging==24.2
    # via
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   matplotlib
    #   nbconvert
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.0.0
    # via
    #   -r requirements/requirements.in
    #   matplotlib
platformdirs==4.3.6
    # via jupyter-core
prometheus-client==0.21.0
    # via jupyter-server
prompt-toolkit==3.0.48
    # via
    #   ipython
    #   jupyter-console
psutil==6.1.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pycparser==2.22
    # via cffi
pyexr==0.4.0
    # via -r requirements/requirements.in
pygments==2.18.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
    #   matplotlib
python-json-logger==2.0.7
    # via jupyter-events
pyyaml==6.0.2
    # via jupyter-events
pyzmq==26.2.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.32.3
    # via jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.21.0
    # via
    #   jsonschema
    #   referencing
ruff==0.8.0
    # via -r requirements/requirements.testing.in
scipy==1.14.1
    # via -r requirements/requirements.in
send2trash==1.8.3
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.4.0
    # via nbconvert
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
tqdm==4.67.1
    # via -r requirements/requirements.in
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.9.0.20241003
    # via arrow
uri-template==1.3.0
    # via jsonschema
urllib3==2.2.3
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==24.11.1
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.8.0
    # via jupyter-server
widgetsnbextension==4.0.13
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
